refactor(src): simplify utils file structure
